Add pre/post clusterize hooks.
Added optional willClusterize: and didClusterize: methods to the
REMarkerClusterDelegate protocol to allow applications to do things like
save and restore the selected annotation when re-clustering occurs.

Note: If the delegate provides a willClusterize: method, annotations are
not automatically deselected within the mapView:regionDidChangeAnimated:
method; the delegate is responsible for ensuring that the correct annotation
is selected.